+++ mappingTool/nodesCrawler.js
@@ -1,52 +1,23 @@
 module.exports = {
     nodesCrawler: async function(mySession, data, crawler, properties, configJson) {
-        var nodeIDParser = require('./nodeIDParser');
-        var ignoreNs = properties.get('namespace-ignore');
+        var dataTypes = require('./dataTypes');
+        var objectFolder = require('./objectFolder');
 
         //console.log(await crawler.read("i=86"));
 
         /*const leaf = await crawler.read();
-    console.log(data);*/
+        console.log(data);*/
 
         switch (data.browseName.toLowerCase()) {
             case 'objects':
-                var obj = {};
-                for (i in data.organizes) {
-                    const ignoreNodeId = nodeIDParser.nodeIDParser(data.organizes[i].nodeId, ignoreNs);
-                    if (!ignoreNodeId) {
-                        const browseSubLev2 = await crawler.read(data.organizes[i].nodeId);
-                        console.log(data.organizes[i]);
-                        if (browseSubLev2.hasOwnProperty('hasComponent')) {
-                            var active = [];
-                            var lazy = [];
-                            var commands = [];
-                            for (k in browseSubLev2.hasComponent) {
-                                const ignoreNodeIdSubLev2 = nodeIDParser.nodeIDParser(
-                                    browseSubLev2.hasComponent[k].nodeId,
-                                    ignoreNs
-                                );
-                                if (!ignoreNodeIdSubLev2) {
-                                    console.log('mmiaComponent');
-                                    console.log(browseSubLev2.hasComponent[k]);
-                                }
-                            }
-
-                            obj[data.organizes[i].browseName] = {
-                                service: properties.get('fiware-service'),
-                                subservice: properties.get('fiware-service-path'),
-                                active: active,
-                                lazy: lazy,
-                                commands: commands
-                            };
-                            configJson.types = obj;
-                        }
-                    }
-                }
                 console.log('Browsing Objects Folder');
+                configJson = await objectFolder.objectFolder(data, configJson, crawler, properties);
                 break;
 
             case 'types':
                 console.log('Browsing Types Folder');
+                const dTypes = await dataTypes.dataTypes(data);
+                console.log(dTypes);
                 break;
 
             case 'views':
@@ -56,7 +27,6 @@ module.exports = {
             default:
                 console.log(`Nothing to be browsed`);
         }
-
         return configJson;
     }
 };
